Corrijo llamado a config_destroy

+++ COM/lqvg/protocolo.h
@@ -15,7 +15,7 @@ typedef struct struInsert{
 	char * nom_tabla;
 	uint16_t key;
 	char * value;
-	unsigned int timestamp;
+	double timestamp;
 } t_insert;
 typedef t_insert* tp_insert;
 
@@ -45,7 +45,7 @@ typedef struct struDrop{
 typedef t_drop* tp_drop;
 
 typedef struct struSelectRta{
-	unsigned int timestamp;
+	double timestamp;
 	uint16_t key;
 	char* value;
 } t_select_rta;
@@ -82,10 +82,10 @@ int prot_recibir_int(int);
 
 void prot_enviar_select(char *, uint16_t, int);
 tp_select prot_recibir_select(int, int);
-void prot_enviar_respuesta_select(char *,uint16_t,unsigned int, int);
+void prot_enviar_respuesta_select(char *,uint16_t,double, int);
 tp_select_rta prot_recibir_respuesta_select(int,int);
 
-void prot_enviar_insert(char *, uint16_t, char *, unsigned int, int);
+void prot_enviar_insert(char *, uint16_t, char *, double, int);
 tp_insert prot_recibir_insert(int, int);
 void prot_enviar_respuesta_insert(int);
 enum MENSAJES prot_recibir_respuesta_insert(int);//@NO necesita antes hacer un recibirCabecera
